Implement SGD and NeuralNet tests.
Implemented unit/integration tests for SGD optimizer and NeuralNet.

SGD is tested by loading a small subset of MNIST data and checking that
after running SGD for 1, 2 and 4 steps the loss is still descreasing.

NeuralNet forward pass is tested by manually combining two linear layers
and then checking that the results match a NeuralNet with two linear
layers. Backward pass is tested in a similar manner.